data_fetcher.py
# data_fetcher.py
import yfinance as yf
from newsapi import NewsApiClient
import requests
from bs4 import BeautifulSoup
import logging
from config import NEWSAPI_KEY # Import API key

# Setup logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')

# Initialize NewsAPI client if key exists
if NEWSAPI_KEY:
    newsapi = NewsApiClient(api_key=NEWSAPI_KEY)
else:
    newsapi = None
    logging.warning("NewsAPI client not initialized because NEWSAPI_KEY is missing.")

# --- Placeholder Functions for Phase 0 ---

def get_stock_info(ticker):
    """
    Fetches basic stock information using yfinance.
    (Phase 1 Implementation Needed)
    """
    logging.info(f"Placeholder: Fetching stock info for {ticker}")
    # --- Implementation for Phase 1 ---
    # try:
    #     stock = yf.Ticker(ticker)
    #     info = stock.info # Fetch basic info
    #     # Fetch current price - more robust methods exist for real-time
    #     hist = stock.history(period="1d")
    #     current_price = hist['Close'].iloc[-1] if not hist.empty else info.get('currentPrice', 'N/A')
    #     logging.info(f"Fetched info for {ticker}")
    #     return {"info": info, "current_price": current_price}
    # except Exception as e:
    #     logging.error(f"Error fetching yfinance data for {ticker}: {e}")
    #     return None
    # -----------------------------------
    print(f"[Placeholder] Would fetch yfinance data for {ticker}")
    return {"info": {"symbol": ticker, "longName": "Placeholder Company Name"}, "current_price": 100.00} # Dummy data

def get_us_news(ticker_or_company_name):
    """
    Fetches news articles for a US stock using NewsAPI.
    (Phase 1 Implementation Needed)
    """
    logging.info(f"Placeholder: Fetching US news for {ticker_or_company_name}")
    if not newsapi:
        logging.error("Cannot fetch US news: NewsAPI client not available.")
        return []
    # --- Implementation for Phase 1 ---
    # try:
    #     # Use q for keyword search, consider adding sources or domains
    #     all_articles = newsapi.get_everything(q=ticker_or_company_name,
    #                                           language='en',
    #                                           sort_by='publishedAt', # or relevancy
    #                                           page_size=20) # Adjust page size as needed
    #     logging.info(f"Fetched {len(all_articles.get('articles', []))} articles for {ticker_or_company_name} from NewsAPI.")
    #     # Return a list of dictionaries with relevant fields
    #     return all_articles.get('articles', [])
    # except Exception as e:
    #     logging.error(f"Error fetching NewsAPI data for {ticker_or_company_name}: {e}")
    #     return []
    # -----------------------------------
    print(f"[Placeholder] Would fetch NewsAPI data for {ticker_or_company_name}")
    # Dummy data
    return [
        {'title': 'Placeholder News Headline 1', 'url': 'http://example.com/news1', 'source': {'name': 'Example Source'}, 'publishedAt': '2023-01-01T12:00:00Z'},
        {'title': 'Placeholder News Headline 2', 'url': 'http://example.com/news2', 'source': {'name': 'Another Source'}, 'publishedAt': '2023-01-01T11:00:00Z'}
    ]

def scrape_indian_news(ticker):
    """
    Placeholder for scraping news for Indian stocks.
    (Phase 2 Implementation Needed using requests/bs4 or playwright)
    """
    logging.info(f"Placeholder: Scraping Indian news for {ticker}")
    # --- Implementation for Phase 2 ---
    # headers = {'User-Agent': 'Mozilla/5.0 ...'} # Add a user agent
    # url = f"https://www.financialnewssite_example.com/search?q={ticker}" # Example URL structure
    # try:
    #     response = requests.get(url, headers=headers, timeout=10)
    #     response.raise_for_status() # Raise error for bad responses (4xx or 5xx)
    #     soup = BeautifulSoup(response.text, 'html.parser')
    #     # Find relevant HTML elements containing headlines, links, etc.
    #     # articles = soup.find_all('div', class_='news-item') # Example selector
    #     # extracted_data = []
    #     # for article in articles:
    #     #     title = article.find('h3').text
    #     #     link = article.find('a')['href']
    #     #     # Extract other data...
    #     #     extracted_data.append({'title': title, 'url': link, ...})
    #     # return extracted_data
    # except requests.exceptions.RequestException as e:
    #     logging.error(f"Error scraping news for {ticker}: {e}")
    #     return []
    # -----------------------------------
    print(f"[Placeholder] Would scrape Indian news for {ticker}")
    return [] # Return empty list for now