requirements.txt
streamlit 
yfinance 
newsapi-python 
requests 
beautifulsoup4 
transformers 
torch 
python-dotenv